# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PampaBooks API
  version: 1.0.0
  description: API para gerenciar Usuários, Avaliações, Pedidos e Catálogo na loja online PampaBooks.
  contact:
    name: Vitor Fiorio
    email: contato@pampabooks.com

servers:
  - url: http://localhost:3000
    description: Servidor do Serviço de Usuários
  - url: http://localhost:3001
    description: Servidor do Serviço de Avaliações
  - url: http://localhost:3002
    description: Servidor do Serviço de Pedidos
  - url: http://localhost:3003
    description: Servidor do Serviço de Catálogo

paths:
  /users/register:
    post:
      summary: Cadastro de usuários
      tags: [Usuários]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                  example: João da Silva
                email:
                  type: string
                  example: joao.silva@gmail.com
                password:
                  type: string
                  example: senha123
      responses:
        201:
          description: Usuário cadastrado com sucesso
        400:
          description: Erro de validação

  /users/login:
    post:
      summary: Login de usuários
      tags: [Usuários]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  example: joao.silva@gmail.com
                password:
                  type: string
                  example: senha123
      responses:
        200:
          description: Usuário autenticado com sucesso
        401:
          description: Credenciais inválidas

  /users/profile:
    get:
      summary: Obter perfil do usuário
      tags: [Usuários]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Perfil do usuário retornado com sucesso
        401:
          description: Não autorizado
    put:
      summary: Atualizar perfil do usuário
      tags: [Usuários]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: João da Silva
                email:
                  type: string
                  example: joao.silva@gmail.com
                password:
                  type: string
                  example: novaSenha123
      responses:
        200:
          description: Perfil atualizado com sucesso
        401:
          description: Não autorizado

  /reviews:
    get:
      summary: Listar avaliações
      tags: [Avaliações]
      responses:
        200:
          description: Lista de avaliações
    post:
      summary: Criar uma avaliação
      tags: [Avaliações]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - livroId
                - userId
                - comentario
                - nota
              properties:
                livroId:
                  type: string
                  example: 644bf3fc1e70dc3b21d2d9d6
                userId:
                  type: string
                  example: 674bf3fc1e70dc3b21d2d9d6
                comentario:
                  type: string
                  example: Excelente livro!
                nota:
                  type: integer
                  minimum: 1
                  maximum: 5
                  example: 5
      responses:
        201:
          description: Avaliação criada com sucesso
        400:
          description: Erro de validação

  /requests:
    get:
      summary: Listar pedidos
      tags: [Pedidos]
      responses:
        200:
          description: Lista de pedidos
    post:
      summary: Criar um pedido
      tags: [Pedidos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - itens
                - total
              properties:
                userId:
                  type: string
                  example: 674bf3fc1e70dc3b21d2d9d6
                itens:
                  type: array
                  items:
                    type: object
                    properties:
                      livroId:
                        type: string
                        example: 644bf3fc1e70dc3b21d2d9d6
                      quantidade:
                        type: integer
                        example: 2
                total:
                  type: number
                  example: 100.50
      responses:
        201:
          description: Pedido criado com sucesso
        400:
          description: Erro de validação

  /catalogs:
    get:
      summary: Listar livros
      tags: [Catálogo]
      responses:
        200:
          description: Lista de livros
    post:
      summary: Adicionar um novo livro ao catálogo
      tags: [Catálogo]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - titulo
                - autor
                - preco
                - estoque
              properties:
                titulo:
                  type: string
                  example: O Senhor dos Anéis
                autor:
                  type: string
                  example: J.R.R. Tolkien
                preco:
                  type: number
                  example: 99.90
                estoque:
                  type: integer
                  example: 100
      responses:
        201:
          description: Livro adicionado ao catálogo
        400:
          description: Erro de validação

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT